+++ packages/core/src/index.ts
@@ -1,5 +1,5 @@
 export { DefaultUserAgent } from "./lib/waku.js";
-export { DefaultPubSubTopic } from "./lib/constants.js";
+export { DefaultPubsubTopic } from "./lib/constants.js";
 export { createEncoder, createDecoder } from "./lib/message/version_0.js";
 export type {
   Encoder,