+++ packages/core/src/lib/constants.ts
@@ -1,4 +1,4 @@
 /**
- * DefaultPubSubTopic is the default gossipsub topic to use for Waku.
+ * DefaultPubsubTopic is the default gossipsub topic to use for Waku.
  */
-export const DefaultPubSubTopic = "/waku/2/default-waku/proto";
+export const DefaultPubsubTopic = "/waku/2/default-waku/proto";