+++ packages/tests/tests/filter/unsubscribe.node.spec.ts
@@ -1,4 +1,4 @@
-import { createDecoder, createEncoder, DefaultPubSubTopic } from "@waku/core";
+import { createDecoder, createEncoder, DefaultPubsubTopic } from "@waku/core";
 import type { IFilterSubscription, LightNode } from "@waku/interfaces";
 import { utf8ToBytes } from "@waku/utils/bytes";
 import { expect } from "chai";
@@ -29,7 +29,7 @@ describe("Waku Filter V2: Unsubscribe", function () {
 
   this.beforeEach(async function () {
     this.timeout(15000);
-    [nwaku, waku] = await runNodes(this, [DefaultPubSubTopic]);
+    [nwaku, waku] = await runNodes(this, [DefaultPubsubTopic]);
     subscription = await waku.filter.createSubscription();
     messageCollector = new MessageCollector();