+++ packages/tests/tests/wait_for_remote_peer.node.spec.ts
@@ -1,4 +1,4 @@
-import { DefaultPubSubTopic, waitForRemotePeer } from "@waku/core";
+import { DefaultPubsubTopic, waitForRemotePeer } from "@waku/core";
 import type { LightNode, RelayNode } from "@waku/interfaces";
 import { Protocols } from "@waku/interfaces";
 import { createLightNode, createRelayNode } from "@waku/sdk";
@@ -40,7 +40,7 @@ describe("Wait for remote peer", function () {
     await waku1.dial(multiAddrWithId);
     await delay(1000);
     await waitForRemotePeer(waku1, [Protocols.Relay]);
-    const peers = waku1.relay.getMeshPeers(DefaultPubSubTopic);
+    const peers = waku1.relay.getMeshPeers(DefaultPubsubTopic);
     const nimPeerId = multiAddrWithId.getPeerId();
 
     expect(nimPeerId).to.not.be.undefined;
@@ -263,7 +263,7 @@ describe("Wait for remote peer", function () {
     await waku1.dial(multiAddrWithId);
     await waitForRemotePeer(waku1);
 
-    const peers = waku1.relay.getMeshPeers(DefaultPubSubTopic);
+    const peers = waku1.relay.getMeshPeers(DefaultPubsubTopic);
 
     const nimPeerId = multiAddrWithId.getPeerId();